Fix improper metadata for rerun episode
In tvmaze.py "lookup by timestamp", there is logic to consider
it a match if the recording is a little bit early. This helps
cases where you set up a rule to record, at say 9:00, and the
broadcaster uses a slightly odd start time, like 9:05. The
code set minTimeDelta to 60 minutes, which allowed such a
match, even if the show started a full hour before the start
time in the TvMaze database.

This minTimeDelta value was way too big. For a 7pm show, it
caused a false match to an 8pm show found in the TvMaze
database. We need to use a smaller default value. Certainly
less than 1 hour. Most TV episodes are either 30 minutes or
60 minutes, so we should set this value smaller than 30
minutes to avoid the same sort of problem with 30 minute
shows.

In this commit, minTimeDelta is now defaulted to 20 minutes.

Refs: #717