Updated fix for improper metadata on rerun episode
In tvmaze.py "lookup by timestamp", there is logic to
consider it a match if the recording is a little bit
early. This helps, for example, if you set up a Manual
Record rule to start at 8:00, but the broadcaster
starts the show at 8:01 or 8:05.

In tvmaze.py, we used to allow an early match for
episodes starting as much as 60 minutes earlier,
but that caused problems.

In this new implementation, we limit the early
match period to, at most, half of the episode
length. This removes the problem of a false
match for a rerun episode followed by a new
episode. If the episode is 60 minutes, then
the maximum early match period is 30 minutes.
If it's 7 minutes, then the maximum early
match period is 3.5 minutes.

Refs: #717